Move the print functions in a new file.
