+++ src/debug.h
@@ -1,0 +1,6 @@
+#ifndef DEBUG_H__
+#define DEBUG_H__
+
+void print_be_tree(const struct config* config, const struct cnode* root);
+
+#endif
